+++ src/CSharp/CodeCracker/CodeCracker.csproj
@@ -69,7 +69,6 @@
     <Compile Include="Refactoring\ChangeAnyToAllCodeFixProvider.cs" />
     <Compile Include="Refactoring\ReplaceWithGetterOnlyAutoPropertyAnalyzer.cs" />
     <Compile Include="Refactoring\ReplaceWithGetterOnlyAutoPropertyCodeFixProvider.cs" />
-    <Compile Include="Refactoring\ReplaceWithGetterOnlyAutoPropertyCodeFixProviderAll.cs" />
     <Compile Include="Refactoring\SplitIntoNestedIfFixAllProvider.cs" />
     <Compile Include="Refactoring\SplitIntoNestedIfCodeFixProvider.cs" />
     <Compile Include="Refactoring\SplitIntoNestedIfAnalyzer.cs" />
